requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-groq
langchain-ollama
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain_huggingface
chromadb
sentence_transformers
faiss-cpu
langchain_chroma
duckdb
pandas
mysql-connector-python

SQLAlchemy